score/requirements.txt
absl-py==1.2.0
anyio==3.6.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.7.0
backcall==0.2.0
black==22.6.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
coinbase==2.1.0
colorama==0.4.4
debugpy==1.6.2
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.4
executing==0.8.3
fastapi==0.79.0
flake8==5.0.4
Flask==2.2.2
flatbuffers==1.12
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h11==0.13.0
h5py==3.7.0
hiredis==2.0.0
httptools==0.4.0
icecream==2.1.3
idna==3.3
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
libclang==14.0.1
limits==1.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mypy-extensions==0.4.3
ndicts==0.1.0
nest-asyncio==1.5.5
nulltype==2.3.1
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
orjson==3.6.8
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
plaid-python==9.9.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycryptodome==3.14.1
pydantic==1.9.1
pyflakes==2.5.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
redis==3.5.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
slowapi==0.1.5
sniffio==1.2.0
SQLAlchemy==1.4.40
stack-data==0.3.0
starlette==0.19.1
style==1.1.0
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.2.0
ujson==5.3.0
update==0.0.1
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1